+++ app/src/landing/LandingPage.tsx
@@ -1,18 +1,17 @@
 import React from 'react';
-import {
-    HeroBlock,
-    StartedBlock,
-    QuoteBlock,
-    ContactsBlock,
-    TechnologiesBlock,
-    ProjectsBlock,
-    BenefitsBlock,
-    ExploreBlock,
-    AskDevelopersBlock,
-    ReleasesBlock,
-    InfoBlock,
-    TeamBlock,
-} from './';
+
+import { TeamBlock } from './TeamBlock';
+import { StartedBlock } from './StartedBlock';
+import { QuoteBlock } from './QuoteBlock';
+import { ContactsBlock } from './ContactsBlock';
+import { TechnologiesBlock } from './TechnologiesBlock';
+import { ProjectsBlock } from './ProjectsBlock';
+import { BenefitsBlock } from './BenefitsBlock';
+import { HeroBlock } from './HeroBlock';
+import { ExploreBlock } from './ExploreBlock';
+import { AskDevelopersBlock } from './AskDevelopersBlock';
+import { InfoBlock } from './InfoBlock';
+import { ReleasesBlock } from './ReleasesBlock';
 import { AppFooter, AppHeader, Page } from '../common';
 import { FlexCell } from '@epam/promo';
 import { analyticsEvents } from '../analyticsEvents';
