+++ app/src/sandbox/skills/components/BigBattery.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import cx from 'classnames';
 import css from './BigBattery.module.scss';
 import { Button, FlexRow, Text } from '@epam/promo';
-import { ISkillLevel } from '../index';
+import { ISkillLevel } from '../utils';
 import { ReactComponent as CrossIcon } from '@epam/assets/icons/common/navigation-close-18.svg';
 
 interface IBigBatteryProps {
