+++ uui-components/src/navigation/MainMenu/MainMenu.tsx
@@ -3,7 +3,7 @@ import {
     IAdaptiveItem, ICanRedirect, IHasCaption, IHasChildren, IHasCX, Link, IHasRawProps, cx, IHasForwardedRef, DropdownBodyProps,
 } from '@epam/uui-core';
 import { BurgerProps } from './Burger/Burger';
-import { MainMenuLogo } from './index';
+import { MainMenuLogo } from '../MainMenu/MainMenuLogo';
 import { AdaptivePanel, AdaptiveItemProps } from '../../layout';
 import { i18n } from '../../i18n';
 import css from './MainMenu.module.scss';
