+++ uui-components/src/pickers/DataPickerRow.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { DataRowProps, uuiMod } from '@epam/uui-core';
-import { FlexRow } from '../index';
+import { FlexRow } from '../layout';
 
 interface DataPickerRowProps<TItem, TId> extends DataRowProps<TItem, TId> {
     renderContent: () => React.ReactNode;
