+++ uui-core/src/data/processing/BaseDataSource.tsx
@@ -1,6 +1,6 @@
 import {
     BaseListViewProps, DataSourceState, IDataSource, IDataSourceView,
-} from '../../types';
+} from '../../types/dataSources';
 
 export abstract class BaseDataSource<TItem, TId, TFilter = any> implements IDataSource<TItem, TId, TFilter> {
     protected views = new Map<any, IDataSourceView<TItem, TId, TFilter>>();
