+++ uui-core/src/data/processing/ListApiCache.tsx
@@ -1,7 +1,7 @@
 import { FAILED, LazyLoadedMap, UNKNOWN } from '../../helpers';
 import { LazyDataSourceApi } from '../../types';
 import { batch } from '../../helpers/batch';
-import { NOT_FOUND_RECORD } from './views';
+import { NOT_FOUND_RECORD } from '../processing/views/tree/ITree';
 
 export interface ListApiSettings<TItem, TId, TFilter> {
     /** Lazy List API used to fetch items */
