+++ uui-core/src/data/validation/index.ts
@@ -1,3 +1,4 @@
-export { validate, blankValidationState, ValidationMode } from './validate';
+export { validate, blankValidationState } from './validate';
+export type { ValidationMode } from './validate';
 export { validateServerErrorState } from './validateServerErrorState';
 export { mergeValidation } from './mergeValidation';
