+++ uui-core/src/services/routing/HistoryAdaptedRouter.tsx
@@ -1,5 +1,7 @@
-import { IRouterContext, Link } from '../../types';
-import { queryToSearch, searchToQuery } from '../../helpers';
+import { Link } from '../../types/objects';
+import { IRouterContext } from '../../types/contexts';
+import { queryToSearch } from '../../helpers/queryToSearch';
+import { searchToQuery } from '../../helpers/searchToQuery';
 
 export interface IHistory4 {
     location: Link;
