+++ uui-core/src/services/routing/NextRouterAdapter.tsx
@@ -1,5 +1,5 @@
 import { IRouterContext, Link } from '../../types';
-import { queryToSearch } from '../../helpers';
+import { queryToSearch } from '../../helpers/queryToSearch';
 
 const createHref = (location: Link, basePath: string) => {
     const { pathname, query } = location;
