+++ uui/components/fileUpload/FileCard.tsx
@@ -4,7 +4,10 @@ import { i18n } from '../../i18n';
 import { cx, FileUploadResponse, formatBytes, IClickable, IHasCX, uuiMod } from '@epam/uui-core';
 import { SvgCircleProgress } from './SvgCircleProgress';
 import { IconContainer } from '@epam/uui-components';
-import { FlexCell, FlexRow, IconButton, Text, Tooltip } from '../../components';
+import { FlexCell, FlexRow } from '../../components/layout';
+import { IconButton } from '../../components/buttons';
+import { Text } from '../../components/typography';
+import { Tooltip } from '../../components/overlays';
 import { fileIcons } from '../../icons/icons';
 import { ReactComponent as RemoveIcon } from '@epam/assets/icons/common/navigation-close-18.svg';
 import { ReactComponent as ErrorIcon } from '../../icons/notification-error-fill-10.svg';
